feat(layout): improved module layout
As there will be plenty of types, it will be better to split it up.
Also learned something about self::<submodule> :).

Insert and and update should be hand-implemented just to see how it's
working. Then there should be some investment to auto-generate this
with `gsl`. Once the latter works ... I could auto-generate all apis,
or adjust the go generator to create rust instead.

Depends on what will be faster ... .