feat(mako): cargo.toml template
It's quite final, and super easy to change and to read.

It seems we want to use namespaces/shared implementations soon to allow
using defs. In our case, we transform the version in a particular way,
which is easy enough, yet I'd like to use it to make the system more
powerful.