fix(discovery): handle discovery urls with $
Some google discovery URLs contain `$discovery` or other variants,
causing the calls to wget to interpret `$d` as an environment variable
instead of a literal. An example is:
`https://logging.googleapis.com/$discovery/rest?version=v2`.

To fix this, the `$` has been escaped so that wget fetches the URL as
expected.