fix(index.html.mako): Finished adjustments to index.html template (fixes #166)
Summary of changes:

- Converted from using span + br tags for formatting to using tables
- Added Bootstrap stylesheet
- Refactored a lot of the logic which was being done in the html ${...}
  tags out into a block which gets run at the start of each api version.
  (hopefully this will make the template easier to maintain in the long
  run)

Possible issue:

I swapped from looping over each key in `tc.keys()` to assuming the keys
will only ever be ["api", "cli"]. This hard codes the keys instead of
getting them dynamically, but makes it easier to format as a table and
lets you pull a lot of the logic out of the template and into a single
block before each table row.

If the types of application in `tc.keys()` ever changes then this
template will need to be updated accordingly!