Fix '-force' target generation for JSON schemata.
In some environments, a variable-existence test in a template was
yielding a false negative which resulted in the template not generating
'-force' targets for API JSON files, even when the `FETCH_APIS`
environment variable was set during template rendering.

This commit adjusts the way the test for the variable presence is done
such that the '-force' targets now render when they should.