Remove old time dependency from API.
Chrono currently depends on an old version of time with a reported
vulnerability: https://rustsec.org/advisories/RUSTSEC-2020-0159

While it does not use any vulnerable code, the dependency may show
up in code vulnerability scans, etc.

This removes the "oldtime" feature from chrono, to remove that.
Also removes the "std" feature because it doesn't seem to be in use
in this code.